more code-cleanups and some bugfixes
